// src/types/resume.d.ts
export type Education = {
  degree: string;
  school: string;
  period: string;
  description: string;
};

export type Experience = {
  position: string;
  company: string;
  period: string;
  location: string;
  description: string;
  technologies: string[];
  responsibilities: string[];
  achievements: string[];
};